+++ Sources/ZcashPaymentURI/ParserContext.swift
@@ -2,7 +2,7 @@
 //  ParserContext.swift
 //  zcash-swift-payment-uri
 //
-//  Created by pacu on 20/03/2025.
+//  Created by pacu on 2025-03-20.
 //
 
 import Foundation
@@ -25,7 +25,6 @@ public enum ParserContext {
 extension ParserContext {
     var sproutPrefix: String {
         switch self {
-            
         case .mainnet:
             "zc"
         case .testnet, .regtest:
@@ -95,7 +94,13 @@ extension ParserContext {
 }
 
 extension ParserContext: AddressValidator {
+    /// checks that address string conforms to ASCII character set, rejects possible Sprout addreses
+    /// and attempts to identify HRPs of the given address match the expected values for the
+    /// corresponding consensus network
     public func isValid(address: String) -> Bool {
+        guard address.conformsToCharacterSet(.ASCIIAlphaNum) else {
+            return false
+        }
         // sprout addresses are not allowed on ZIP-320
         guard !self.isSprout(address: address) else { return false }
         
@@ -103,14 +108,26 @@ extension ParserContext: AddressValidator {
         self.isShielded(address: address)
     }
     
+    /// Tentatively checks HPR prefixes of the given address string for conformance to the
+    /// corresponding network parameter
+    /// - Important: does not check
     public func isTransparent(address: String) -> Bool {
-        address.hasPrefix(self.p2pkhPrefix) ||
+        guard address.conformsToCharacterSet(.ASCIIAlphaNum) else {
+            return false
+        }
+        
+        return address.hasPrefix(self.p2pkhPrefix) ||
         address.hasPrefix(self.p2shPrefix) ||
         address.hasPrefix(self.texPrefix)
     }
     
+    /// Tentatively checks that the given `address` has the expected HRP prefix for the
+    /// correposnding network
     public func isSprout(address: String) -> Bool {
-        address.hasPrefix(self.sproutPrefix)
+        
+        // Naïve checking Testnet Sprout addresses makes 'zt' prefix collide with
+        // 'ztestsapling'
+        address.hasPrefix(self.sproutPrefix) && !address.hasPrefix("ztestsapling")
     }
     
     public func isShielded(address: String) -> Bool {